requirements.txt
distro==1.5.0
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tabula-py==2.2.0
